requirements.txt
ccxt==4.5.40
pandas==2.3.2
pandas_ta==0.4.71b0
python-telegram-bot[job-queue]==22.6
python-dotenv==1.2.2
pytz==2025.2
numpy==2.2.6
matplotlib>=3.8,<4.0
fastapi==0.115.0
jinja2==3.1.4
uvicorn==0.30.6
python-multipart==0.0.20
